create class password salt.
