+++ src/main/java/ffc.airsync.api/dao/MongoOrgDao.kt
@@ -2,6 +2,7 @@ package ffc.airsync.api.dao
 
 import com.mongodb.BasicDBObject
 import com.mongodb.client.FindIterable
+import ffc.airsync.api.dao.PasswordSalt.getPass
 import ffc.airsync.api.printDebug
 import ffc.entity.Organization
 import ffc.entity.User
@@ -15,40 +16,12 @@ import javax.ws.rs.NotFoundException
 
 class MongoOrgDao(host: String, port: Int, databaseName: String, collection: String) : OrgDao, UserDao, MongoAbsConnect(host, port, databaseName, collection) {
 
-    private var SALT_PASS = """
-uxF3Ocv5eg4BoQBK9MmR
-rwPARiCL9ovpr3zmlJlj
-kIQnpzRIgEh8WLFNHyy1
-ALqs9ES1aQlsc47DlG5f
-SbAOMWzMd1T03dyigoHR
-7hox2nDJ7tMJRHab5gsy
-Ux2VxiCIvJtfPAobOxYW
-HazJzQEGdXpmeM2aK6MD
-mpOARM2427A6CY14uomK
-Cxe9aEkJEFtlLLo6NaNW
-yLkbHUfMNDwWeu2BRXuS
-m7BHwYSyKGFJdLnq4jJd
-sr4QI6aK7g3GCm8vG6Pd
-RAtlJZFto0bi9OZta5b4
-DLrNTZXXtB3Ci17sepXU
-HSYUuw11GJmeuiLKgJYZ
-PCHuw2hpoozErKVxEv86
-f6zMttthJyQnrDBHGhma
-j1nrasD5fg9NxuwkdJq8
-ytF2v69RwtGYf7C6ygwD
-"""
-
-    init {
-        val salt = System.getenv("FFC_SALT")
-        if (salt != null) SALT_PASS = salt
-    }
-
     override fun insert(organization: Organization): Organization {
         printDebug("Call mongo insert organization")
         `ตรวจสอบเงื่อนไขการลงทะเบียน Org`(organization)
 
         organization.users.forEach {
-            it.password = getPass(it.password, SALT_PASS)
+            it.password = getPass(it.password)
         }
 
         val genId = ObjectId()
@@ -214,11 +187,10 @@ ytF2v69RwtGYf7C6ygwD
         if (haveUserInDb(orgId, user)) throw BadRequestException("มีการเพิ่มผู้ใช้ ${userInsert.name} ซ้ำ")
 
         val query = Document("id", orgId)
-        userInsert.password = getPass(userInsert.password, SALT_PASS)
+        userInsert.password = getPass(userInsert.password)
         val userDoc = Document.parse(userInsert.toJson())
         val userStruct = Document("users", userDoc)
         val userPush = Document("\$push", userStruct)
-
 
         printDebug("\tCreate user in mongo")
         coll2.updateOne(query, userPush)
@@ -267,7 +239,7 @@ ytF2v69RwtGYf7C6ygwD
         val org = orgDoc.toJson().parseTo<Organization>()
         printDebug("\torg = $org")
 
-        val passwordSalt = getPass(pass, SALT_PASS)
+        val passwordSalt = getPass(pass)
         printDebug("Salt Pass = $passwordSalt")
 
         val user = org.users.find {
