+++ src/main/java/ffc.airsync.api/dao/MongoUserDao.kt
@@ -1,5 +1,6 @@
 package ffc.airsync.api.dao
 
+import ffc.airsync.api.dao.PasswordSalt.getPass
 import ffc.airsync.api.dao.UserDao.Companion.checkBlockUser
 import ffc.airsync.api.printDebug
 import ffc.entity.User
@@ -9,34 +10,6 @@ import org.bson.Document
 import javax.ws.rs.NotFoundException
 
 class MongoUserDao(host: String, port: Int, databaseName: String, collection: String) : UserDao, MongoAbsConnect(host, port, databaseName, collection) {
-
-    private var SALT_PASS = """
-uxF3Ocv5eg4BoQBK9MmR
-rwPARiCL9ovpr3zmlJlj
-kIQnpzRIgEh8WLFNHyy1
-ALqs9ES1aQlsc47DlG5f
-SbAOMWzMd1T03dyigoHR
-7hox2nDJ7tMJRHab5gsy
-Ux2VxiCIvJtfPAobOxYW
-HazJzQEGdXpmeM2aK6MD
-mpOARM2427A6CY14uomK
-Cxe9aEkJEFtlLLo6NaNW
-yLkbHUfMNDwWeu2BRXuS
-m7BHwYSyKGFJdLnq4jJd
-sr4QI6aK7g3GCm8vG6Pd
-RAtlJZFto0bi9OZta5b4
-DLrNTZXXtB3Ci17sepXU
-HSYUuw11GJmeuiLKgJYZ
-PCHuw2hpoozErKVxEv86
-f6zMttthJyQnrDBHGhma
-j1nrasD5fg9NxuwkdJq8
-ytF2v69RwtGYf7C6ygwD
-"""
-
-    init {
-        val salt = System.getenv("FFC_SALT")
-        if (salt != null) SALT_PASS = salt
-    }
 
     override fun insertUser(user: User, orgId: String) {
 
@@ -86,8 +59,7 @@ ytF2v69RwtGYf7C6ygwD
         checkBlockUser(name)
 
         var userDoc: Document? = null
-
-        val query = Document("orgId", orgId).append("name", name).append("pass", getPass(pass, SALT_PASS))
+        val query = Document("orgId", orgId).append("name", name).append("pass", getPass(pass))
 
         mongoSafe(object : MongoSafeRun {
             override fun run() {
